requirements.txt
asgiref==3.2.10
attrs==19.3.0
Automat==20.2.0
beautifulsoup4==4.9.1
bs4==0.0.1
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
constantly==15.1.0
cryptography==3.0
cssselect==1.1.0
Cython==0.29.21
Django==3.0.8
django-crispy-forms==1.9.2
djangorestframework==3.11.0
easy-setup==0.0.1
fake-useragent==0.1.11
gunicorn==20.0.4
html5lib==1.1
hyperlink==19.0.0
idna==2.10
incremental==17.5.0
itemadapter==0.1.0
libxml2dom==0.5
lxml @ file:///C:/Users/Drew/Downloads/lxml-4.5.1-cp38-cp38-win32.whl
numpy==1.19.1
pandas==1.0.5
parsel==1.6.0
Protego==0.1.16
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDispatcher==2.0.5
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
python-dateutil==2.8.1
pytz==2020.1
queuelib==1.5.0
requests==2.24.0
Scrapy==2.2.1
scrapy-selenium==0.0.7
selenium==3.141.0
service-identity==18.1.0
simplejson==3.17.2
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
tqdm==4.47.0
Twisted==20.3.0
Unidecode==1.1.1
urllib3==1.25.9
w3lib==1.22.0
webencodings==0.5.1
xlrd==1.2.0
zope.interface==5.1.0